优化代码逻辑，将密码方法作为单独类管理
