+++ app/src/main/java/dan/fatzero/mypassword/MyPwdMainActivity.java
@@ -28,7 +28,7 @@ import java.util.Random;
 
 public class MyPwdMainActivity extends AppCompatActivity {
 
-
+private GeneratePassword generatePassword;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -48,7 +48,7 @@ public class MyPwdMainActivity extends AppCompatActivity {
         setSupportActionBar(materialToolbar);
 
 
-        GeneratePwd generatePwd = new GeneratePwd();
+        generatePassword = new GeneratePassword();
 
         //生成密码
         generate_button.setOnClickListener(view -> {
@@ -95,7 +95,8 @@ public class MyPwdMainActivity extends AppCompatActivity {
             }
 
             String[] args = {pwdArg1,pwdArg2};
-            String pwdRaw = generatePwd.main(1,pwdLen_num,pwdSeed_num,pwdCaps,pwdSpeChar,args);
+            String pwdRaw = generatePassword.main(1,pwdLen_num,pwdSeed_num,pwdCaps,pwdSpeChar,args);
+           //String pwdRaw = generatePwd.main(1,pwdLen_num,pwdSeed_num,pwdCaps,pwdSpeChar,args);
             generate_pwd.setText(pwdRaw);
             Toast.makeText(MyPwdMainActivity.this, "密码已生成", Toast.LENGTH_SHORT).show();
         });
@@ -128,109 +129,6 @@ public class MyPwdMainActivity extends AppCompatActivity {
 
     }
 
-    //生成密码类
-    private class GeneratePwd{
-
-
-        public String main(int alg_id,int stringLen,int seed,int setCaps,int setSpeChar,String[] args){
-            String alg_result = "";
-            switch (alg_id){
-                case 1://Abby算法
-                    alg_result = algAbby(seed,args);
-                    break;
-                default:
-            }
-
-            //截取为一定长度的密码
-            String pwdRaw = StringUtils.substring(alg_result,0,stringLen);
-
-            if(setCaps == 1){
-                pwdRaw = settingCaps(seed,pwdRaw);
-            }
-
-            if(setSpeChar == 1){
-                pwdRaw = settingSpeChar(seed,pwdRaw);
-            }
-            return pwdRaw;
-        }
-
-        // 密码算法，算法昵称：Abby，id=1
-        public String algAbby (int seed,String[] args){
-            String random_string=randomSeed(seed);
-            //Toast.makeText(MyPwdMainActivity.this,random_string,Toast.LENGTH_SHORT).show();
-            String combineArgs="";
-            for(int i = 0;i < args.length;i=i+1){
-                combineArgs = combineArgs + args[i];
-                Log.d("MyP2","arg:" + args[i]);
-            }
-            Log.d("MyP2","combineArgs:"+combineArgs);
-            String pwdRaw = combineArgs+random_string;
-            String md5Str = getMD5Str(pwdRaw);
-
-            return md5Str;
-        }
-
-
-        //String转MD5
-        public  String getMD5Str(String str) {
-            byte[] digest = null;
-            try {
-                MessageDigest md5 = MessageDigest.getInstance("md5");
-                digest  = md5.digest(str.getBytes("utf-8"));
-            } catch (NoSuchAlgorithmException e) {
-                e.printStackTrace();
-            } catch (UnsupportedEncodingException e) {
-                e.printStackTrace();
-            }
-            //16是表示转换为16进制数
-            String md5Str = new BigInteger(1, digest).toString(16);
-            return md5Str;
-        }
-        // 随机数
-        public String randomSeed(int num){
-            Random r = new Random(num);
-            String random_string = String.valueOf(r.nextInt(10000));
-            return random_string;
-        }
-
-        //替换为大写字母
-        public String settingCaps(int seed,String rawString){
-
-            String s = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
-            char[] c = s.toCharArray();
-            for(int i = 0;i<Math.ceil(rawString.length()*0.1);i=i+1){
-
-                Random r = new Random(seed);
-                String cap = String.valueOf(c[r.nextInt(c.length)]);
-                int rep_index = r.nextInt(rawString.length());
-                StringBuilder sb = new StringBuilder(rawString);
-                rawString = String.valueOf(sb.replace(rep_index,rep_index+1,cap));
-                seed = seed +1;
-            }
-
-            String capsString = rawString;
-            return capsString;
-        }
-
-        public String settingSpeChar(int seed,String rawString){
-
-            String s = "!@#$%^&*_";
-            char[] c = s.toCharArray();
-            for(int i = 0;i<Math.ceil(rawString.length()*0.1);i=i+1){
-
-                Random r = new Random(seed*2);
-                String cap = String.valueOf(c[r.nextInt(c.length)]);
-                int rep_index = r.nextInt(rawString.length());
-                StringBuilder sb = new StringBuilder(rawString);
-                rawString = String.valueOf(sb.replace(rep_index,rep_index+1,cap));
-                seed = seed +1;
-            }
-
-            String specCharString = rawString;
-            return specCharString;
-        }
-
-    }
 
 
     public boolean onCreateOptionsMenu(Menu menu){
